requirements.txt
python-telegram-bot>=21.0
telethon>=1.36.0
aiosqlite>=0.20.0
python-dotenv>=1.0.0
cryptography>=42.0.0
